+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Added playback stalled feature in iOS. iOS version should behave same as Android once video failed to load.
 * Added BetterPlayerTheme to controls configuration (added by https://github.com/maine98).
 * [BREAKING_CHANGE] Changed custom controls builder in BetterPlayerControlsConfiguration. Now it accepts BetterPlayerController.
-
+* Exposed BetterPlayerPlaylistState and betterPlayerController getter within.
 
 ## 0.0.43
 * Added autoDispose flag in BetterPlayerConfiguration